+++ model_fetch/lib/model_fetch.dart
@@ -1,3 +1,5 @@
 export 'src/abstract_client_filter.dart';
 export 'src/abstract_filter.dart';
 export 'src/lazy_load_bloc.dart';
+export 'src/load_status.dart';
+export 'src/model_by_id_bloc.dart';