+++ model_fetch/lib/src/load_status.dart
@@ -1,0 +1,6 @@
+enum LoadStatus {
+  notTried,
+  loading,
+  ok,
+  error,
+}