+++ model_fetch_firestore/lib/model_fetch_firestore.dart
@@ -1,3 +1,4 @@
-export 'src/firestore_lazy_load_bloc.dart';
-export 'src/firestore_frozen_lazy_load_bloc.dart';
+export 'src/lazy_load/firestore_lazy_load_bloc.dart';
+export 'src/lazy_load/firestore_frozen_lazy_load_bloc.dart';
+export 'src/model_by_id/firestore_live_by_id_bloc.dart';
 export 'src/query_builder.dart';