+++ model_fetch_firestore/lib/src/model_by_id/firestore_live_by_id_bloc.dart
@@ -1,0 +1,30 @@
+import 'package:cloud_firestore/cloud_firestore.dart';
+
+import 'package:model_fetch/model_fetch.dart';
+import 'package:model_interfaces/model_interfaces.dart';
+
+class FirestoreLiveByIdBloc<T extends WithId<String>>
+    extends ModelByIdBloc<String, T> {
+  final CollectionReference<T> collectionReference;
+
+  final DocumentReference<T> _doc;
+
+  FirestoreLiveByIdBloc({
+    required this.collectionReference,
+    required String id,
+  })  : _doc = collectionReference.doc(id),
+        super(id: id) {
+    _doc.snapshots().listen(_onModelChanged);
+  }
+
+  void _onModelChanged(DocumentSnapshot<T> documentSnapshot) {
+    final model = documentSnapshot.data();
+
+    emitStateIfChanged(
+      ModelByIdState(
+        model: model,
+        status: model == null ? LoadStatus.error : LoadStatus.ok,
+      ),
+    );
+  }
+}