+++ model_fetch/CHANGELOG.md
@@ -1,3 +1,8 @@
+## 0.2.0
+
+* **BREAKING**: Renamed `LazyLoadStatus` to `LoadStatus`.
+* Added `ModelByIdBloc`.
+
 ## 0.1.0
 
 * Initial release.