+++ model_fetch_firestore/CHANGELOG.md
@@ -1,3 +1,12 @@
+## 0.2.0
+
+* Added `FirestoreLiveByIdBloc`.
+
+## 0.1.1
+
+* Removed debug output and comments.
+* Support client filters.
+
 ## 0.1.0
 
 * Initial release.